WIP - split out linkdata provider from hook
